requirements.txt
pymssql==2.2.2
azure-batch==10.0.0
azure-storage-blob==12.8.1
petl==1.7.4
pytest==6.2.4
azure-storage-file-datalake==12.4.0
azure-datalake-store==0.0.52
azure-mgmt-batch==16.0.0
pandas==1.3.0
#pylint==2.9.5
#black==21.7b0
